Remove the "init_timer" and similar emulations, standardize on "setup_timer".
With the drbd-8.4 commit 2b11dfeb06e73242304a13b838de505f6982b860 this
is now the easiest thing to do, and it actually fixes crashes; before
this change, the DPC was initialized with a timer->function==NULL value.

After this change, we just call the common callback function, which then
runs whatever is _now_ defined to be run.