RtlRandomEx() is incompatible with the Driver Verifier.
Which is bad, because basically we'd like to use both... but as the
Driver Verifier brings more value, we ditch RtlRandomEx() and do our
own randomness. Like so many other people, too...

Luckily the callers of that are rather limited in nature - we don't
need any cryptography quality stuff here.