printk_to_syslog: Print all messages to Windows Debugging facility.
These messages can be seen by DbgView.exe tool on the windows
host. Unfortunately we cannot send all messages to the syslog
host, because Send() may sleep (which we must not when the system
is within an IRQ, much like it is on Linux). By printing all messages
to the local debugging facility, we have a consistent log which we
can work with (unless there is a blue screen of course).