Initialize max_hw_sects in queue.
Now resync requests are sent in chunks of (up to) 1M which makes
resync much faster (was 4K chunks, because max_hw_sects was 0).

Right now, resync between Windows and Linux is still a bit slower
(factor 4) than beween Linuxes, this is maybe because the I/O
stack is slower under Windows (?).