TODO update: Sync Performance Problem solved.
Reason was that we always waited until Windows calls the completion
routine, which some Windows versions do (rightfully) only after
TCP/IP ACK is received from the peer. That way we had an effective
Window size of 4KB which in turn resulted in a network throughput
of about 40KB/sec on our setup. Now we reach the maximum of
100MBit/sec (when Windows 10 box is SyncSource), and about
40MBit/sec in the other direction (when Windows 7 VM is SyncSource),
the latter is probably because it is a VM.