Fix MSG_DONTWAIT behaviour for recv().
It is not directly supported but must not block. This patch
fixes the case where connection state was stuck in NetworkFailure
(because receiver wasn't correctly terminated).

This patch also contains a force_sig to the parent thread
(SIGCHLD) which we probably want to remove again (along with
the parent relationship?)