windrbd_device: Only complete IRPs when all requests have finished.
Also if there is an error. We had the case that IO manager keeps
repeating the I/O request (but this might also have been follow
up requests). BSOD on failure of > 1Meg requests seems now be
fixed (need to cleanup code). We, however had a BSOD on out of
memory (but this is a different thing).