Install userland binaries to C:\windrbd\usr\sbin
We want a well known path for the binaries. Somehow it does
not work to have the service binary (windrbdsvc.exe) in
Program Files or in system32.

For now, all userland binaries live in C:\windrbd\usr\sbin
and the path of the usermode helper is hardcoded in the
DRBD driver (we cannot change it via an ioctl yet, under
Linux this is done via module parameters, which we don't
have under Windows).