Reworked timeouts in user mode helpers.
We have 2 timeouts now: one until daemon fetches the request
(1 second), if not chances are high that the daemon is not
running. Timeout 2 is for the script to run and the daemon
returning the exit status of the script. This may take longer
because we are now using PowerShell, which is extremly slow,
even when it just returns an exit status.