Fixed propagation of DRBD filesystem flag to peers.
Problem was that bi_first_element is changed by generic_make_request()
and always was > 0 by the time the request is completed. We now
disregard the value of bi_first_element and use an atomic flag
to check if the boot sector was already patched.

The problem most likely came in when the max_mdl_elements constant
was finally set to 1 (right after beta4), the race was there before
but never happened since before that bi_first_element was always 0.

Also fixed a ManTech implementation of test_and_change_bit (it
always assumed IRQL being passive).