Fixed BSOD on connect (Windows Server 2016)
This patch resolves a race condition which was introduced
when kthread_create() was redesigned. The race caused
task->windows_task being invalid sometimes, which was used
at the beginning of the ack_receiver thread
(sched_setscheduler()).

As this test shows, fixing this race caused the BSOD to disappear:

drbdadm up w0
disconnect-connect-loop.sh

It might also be the reason of other BSOD observed lately (see
TODO).