No extra copies in upper device on read().
Works now, if backing device is NTFS there is a Access denied
error when accessing the 'protected' sectors of NTFS (as expected).

Note we still need the hide-filesystem because of those access
denied thing (drbdmeta for internal meta data also fails).

No BSOD so far when accessing the NTFS backing device.