For now, do not unlock pages after MDL creation.
This sounded like a bad idea, we needed it when we used
MmBuildMdlForNonPagedPool (in which case pages must not be
locked), but we are not using MmBuildMdlForNonPagedPool any
more.

Also commented out loop for additional MDL elements, this is
not needed any more since there are drivers (Windows 10 USB
stick) which can handle only one MDL element (and BSOD if
there are more ...).