Printk's on boot should work now.
We see them only once the net interface is configured, which is logical.
The problem is probably a logic bug in SendTo (and also Send): it
does not detect when the network is working again after an error.
(socket->error_status isn't reset). We for now open a new socket
if the socket goes into an error status.