Removed check for MSG_NOSIGNAL
It was wrong at this place. MSG_NOSIGNAL means that no signal should
be *sent* on connection loss (SIGPIPE), which we don't do anyways.
It has nothing to do with receiving (handling) signals sent from
somewhere else. This last commit re-introduced the stuck in
NetworkFailure bug (which happens when a signal is not handled,
DRBD uses signals to abort recv() operations).