Features: Updated time table.
And moved system volume before auto promote. Reasons are that
we don't need, for a proof of concept, auto promote for the
system volume feature, implementing auto-promote requires
solving complex problems (we need to create the device on
new-minor, but if we do we need to find out how to invalidate
Windows file system cache) and we actually have a beta customer
that needs system volume (but not auto promote) feature.