(Manual) Patches to DRBD drbd_transport_tcp.c
Remember when manually patching new files of DRBD the base
must be the result of a plain transform run (which runs the
cocci, perl and shell scripts on the file, run make on a
Linux box to generate the file). Based on that
the patch files are applied. It works now (thanks to git
rebase) and creates a drbd_transport_tcp.c which is similar
(while not exactly the same) as the original drbd_transport_wtcp.c

The changes are mostly last minute patches like elimiation of
the seperate connecttimeo, skipping patching the write same
check and removing the timeo_ret parameter to wait_event_interruptible.