Close incoming wsk sockets that are not accepted.
It happens frequently that sockets are "accepted" on the
WSK layer (incoming_connection callback is called) but
they are not needed any more by DRBD (because it already
has 2 open connections) and DRBD never calls kernel_accept()
on them. Close that WSK sockets when the listening socket
containing them is closed, else a shutdown of the WSK
layer on driver unload will stall (and also a later windows
shutdown).

This patch also changes the accept socket spinlock to a
mutex (because close might be called inside the mutex).