Log kmalloc errors.
Currently, printk does a kmalloc internally, which causes
kmalloc not to print the message when it fails. We now
have a counter and hopefully will see kmalloc failures
after they happen (eventually printk works again, but is
seems that we loose messages, cstate is NetworkFailure with
printk guarding them but we don't see those messages).