receive_a_lot test thread (and blocking accept).
This (partially temporary) patch adds a thread that listens
on port 5678 and then receives until connection is aborted.
Its purpose is to test if there is a resource leak (causing
Win7 hang) in the receive path of the driver.

This patch also adds support for blocking kernel_accept()
until a connection comes in.