Fixed wait_event_interruptible_timeout macro.
When using same variable for timeout and return value (which
is actually quite common), the variable was set to 0 before
anything else happens. So, in that case, timeout was always
0, which made certain things (amogst others, the network
connect code and auto promote). Network has been fixed manually
before.

Furthermore, local variables which shadow automatic variables
have been renamed with a __x suffix (in some cases, t was used
for timeout but also by the macro internally).

Does this have to be a macro at all? We should convert it to
functions and implement them properly (i.e. at least without
busy loop) after 0.9.2.