load unload test: now running with connected resource again.
Currently testing .. it ran 77 times without BSOD on my Win7
VM. It seems that implementing reap_all_threads() solved the
BSOD on unload. Also we had a (temporary) BSOD on disconnect
which was probably caused by extensive use of printk's in
the DRBD code. We now re-made DRBD from the patches and it
seems to be gone.