Reverted wait_for_sendbuf function.
Altough I can't see the reason at the moment this seems to fix the
BSOD we observed on running a disconnect-connect-waitconnect
loop after approx. 30 iterations (peer is Linux and up).

Eventually we want to fix the my_wait_interruptible() function
(and also use it in other places, like receive or connect).