Deleting multicast group when file used for creating it is closed
This should fix the wait-connect (and other wait commands) bug
that caused problems (sequence number mismatch).

We record the file object in the multicast group (it is passed
via stack location), when windrbd_close() is called by the
IO manager later on that file, we remove that multicast group
again. This also works when the process that created the
multicast group exits without explicitly closing the file,
so there are no changes neccessary to drbd-utils.