Fixed rcu_lock implementation
It is really a rw lock (with rcu_read_lock() just locking
for reading (others may read as well) and call_rcu and
synchronize rcu being write locks (which must not interfere
with read and other write locks).

System hangup on drbdadm up is fixed now (was introduced
with the last 2 commits).