boot device: Start early to create boot device.
We accidentially tried this with C: as the mount point
of the boot device. This lead us to an unbootable system
because even tough C: existed as a disk, this was
redirected to the DRBD boot device. The boot device
failed to connect (sock_create_kern failed, err = -22)
because network was not initialized yet. We should
wait for network init (via a NotificationEvent) before
attempting to connect the device (however this may
fail for the diskless case, because network needs
system root device? Not clear, give it a try).