Mount as drive C:
This prevents the dev VM to boot, since the WinDRBD device
overrules the real (virtual) harddisk. This shows that the
concept of waiting for becoming primary basically works
however currently Windows needs something else from the
System Volume to make networking work. Currently on booting
the dev VM, it deadlocks after loading the drivers (the
NIC driver is loaded). Maybe we need to tell Windows somehow
that we are network booting.

Also try this with the diskless client VM, maybe it already
works there.