Load order group not Primary Disk
The problem is that WinDRBD overwrites a local attached disk,
which leads to unbootable dev VM which is currently not
what we want. Maybe loading the WinDRBD driver a little
bit later won't overwrite the C: system volume.

Update: Also with default Load Order Group, the C: drive
is overwritten, however it succeeds to show the GUI
with the WinDRBD device as boot device for 1-2 minutes
then stalls. Now testing with Diskless Client.