Mount boot dev while still in DriverEntry.
We had the case that Windows fails to find the boot device.
It BSODs in less that one second after DriverEntry is called.
So this is probably a race condition with Windows trying to
open C: when it does not exist yet.

Also changed (for testing) drive letter to W: (need to set to
C: later again).