Call IoReportRootDevice to tell PnP manager that we're there.
Now, AddDevice is called and the DRBD "bus" device is created.
It also gets Plug and play requests (0x18, 0x0d, 0x00 and 0x020),
we need to implement them and add devices to the bus. The device
also shows up in the device manager as Unknown device under
Other devices.