Not calling report root device.
When installed via Device Manager the bus device is hopefully
created for us.

Update: Yes it is. AddDevice is called on the device we created
via device manager. Also bus relations are queried correctly.
We now add DRBD devices to the bus. BSOD, however PnP manager
previously tried to remove the bus device. The reason is
yet unknown (properties of the device showed the WinDRBD driver
as valid driver), probably we need to implement some other
QUERY_DEVICE PnP minors.