Have seperate INF file for the disk devices.
It seems that this is the way to go. On drbdadm up an Unknown
Device appeared, then I did Update drivers and selected the
directory the inf files resided in and Windows seems to have
automatically selected the correct inf file (windrbd-disk.inf.in)
The main difference to the old inf file is that this is
of Class DiskDrive while the windrbd.inf file is (intentionally)
of Class ScsiController.

We will soon see if further DRBD devices are recognized correctly.
Also note that Windows automatically added partitionmanager
(PARTMGR.SYS) as filter driver for the WinDRBD device, which
is good since we will need this feature soon for the boot
feature.

Currently Windows still (also after a reboot) reports that some
resources are not available. This is probably because we still
don't report the number of devices on the bus correctly and/or
fail to report properties of the disk device correctly (some
things are yet unimplemented).