Hardware ID for disks is now WinDRBDDisk
Without any minors. On drbdadm up <res> a new device is now
automatically associated with the WinDRBD driver (one has to
once do update drivers in the device manager first). So
almost everything looks good, except for the resource
thing (which is probably because we don't already implement
all PNP protocol requests on the bus and also on the disk
side).

Note that AddDevice currently fails to attach the disk objects
on the bus, which might be the reason for the resource
error.