Always return the Windows device for TargetDeviceRelations.
And not the device stacked over it. It seems that we don't
really need the extra logic in AddDevice (adding more
device objects atop of the Windows device). This should
be reverted one day.

Disk devices are now reported as working, however partitions
cannot be detected. Also a new blue screen on closing the
device.