Re-enabled creating boot device.
Right now it waits forever for the network in boot thread and
also waits forever in START_DEVICE for the device to become
primary. Network is initialized after pnp START_DEVICE is finished,
which never happens. Probably because we are setting the
load order group to Primary Disk, which causes windrbd pnp
devices being initialized before the network (which is the
wrong way).