More debugging on I/O.
We still don't know what Windows is doing when opening the
WinDRBD device on boot. Most likely this is some kind of
I/O which we should see now.

Update: It seems that Windows just opens the file without
doing any I/O or ioctl's and closes it, then it deadlocks.