Do not print kernel log via UDP.
There might be a bug which causes the printk fail while
printing via UDP.

Update: Yes. Sending a UDP packet at the wrong point in
time crashes the pnp thread (I suppose in the completion
routine of the printk, since we observe a crashing memcpy).

What we have now is Windows waiting in windrbd_create() for
becoming Primary (which never happens in our copy VM since
there the network does not work).