howto update: how to enable driver verifier
We actually found a mutex_lock with raised IRQL (DISPATCH_LEVEL)
in wsk_incoming_connection (it is sometimes, on boot, called with
IRQL at DISPATCH_LEVEL in which sleeping is not allowed). Next
patch will fix this. Then, if we are lucky this will also solve
the boot lock we observe currently when netbooting via WinDRBD.