winsocket layer: Replaced accept mutex by spinlock.
The reason is that wait_for_incoming_connection() is (sometimes)
called with IRQL at DISPATCH_LEVEL in which case we must not
sleep. Driver verifier found this bug, which maybe also is
responsible for the boot hang.

Update: didn't solve the boot hang but it was a good idea anyway
to fix this.