Implemented most of the Windows specific SCSI helper functions.
Except SRB_FUNCTION_EXECUTE_SCSI, which does the actual SCSI
OPs. We are copiing more or less from ATA over Ethernet (GPL
version 3), since we are also GPL there should be no problem
with that. See http://winaoe.org for more details on aoe
driver. Thanks a lot to the anonymous author ("V.") of this
driver.