Not dereferencing device object on REMOVE SRB request.
We found that all of a sudden our DRBD Windows device
(\Device\Drbd5) disappears together with the drive letter
link. Either this is a new bug or it has been around
since we successfully stacked the disk.sys driver over
windrbd.sys. In the latter case the reason might be that
Windows still complains that it can't find a driver for
the device (and the PnP manager removes the device).