Report device object for SCSI requests.
We are not sure if SCSI requests are sent to the disk or
the bus object. Currently we get a BSOD on removing the
windrbd disk object when it is deleted via PnP when installing
the driver. The reason might be that the windrbd disk object
is already deleted by some other magic.