Implemented IRP_MN_QUERY_PNP_DEVICE_STATE
This seems to have solved the Could not start device issue
that shows up in device manager for the Disk devices.

Right now, we cannot stack the Disk driver over the WinDRBD
driver, it used to work but for some reason we either see
WinDRBD and Partition manager or Disk driver and Partition manager.