Implemented IOCTL_STORAGE_QUERY_PROPERTY and IOCTL_SCSI_GET_ADDRESS ioctl's
Those are neccessary for the Device object being treated as a
SCSI disk device. We now get various SCSI OPs sent (25, 28,
1a, ...). This is what we were looking for the past week.

To summarize: the loader from AoE probably helps it is unclear
if we need it.

The drivers stacked over the WinDRBD disk object are: disk.sys
as a function driver, partmgr.sys as a filter driver. WinDRBD
itself is just the bus driver (the Virtual SCSI Bus Device
has to be created manually).

The QUERY_ID PnP request must report the device as GenDisk compatible
(and nothing else).

Creating an extra device object above the one created in
new-minor is useless (that is AddDevice should only add the
bus object).

Forwarding unknown PnP requests probably helps not sure if
it is really neccessary.

We now proceed to implement the various SCSI ops (READ/WRITE/
SENSE/...)