SCSI: Wait until size is valid in GET_CAPACITY
On receiving SCSIOP_READ_CAPACITY16 we need to wait until
we are either connected to a diskful peer or attached to
a local disk. There is an additional patch to drbd_nl.c
which we will add in a followup commit (once we've tested
if it works).

What this patch is missing is unsetting the event in case
we get detached (and size drops to zero). We'll add that
soon.

This patch currently BSOD's on drbdadm up (when setting
the event in blkdev_get_by_path on attach).