Only touch capacity event in drbd_set_my_capacity.
This is done via a DRBD patch. Do not set the event in
blkdev_get_by_path() in this function, the block_device
is a different object than the one used in the SCSI device
requests (the latter is the upper device object that has
a valid windows_device associated with it). In blkdev_get_by_path
the block_device object is the object for the attached
device (the lower device object).

What is missing is support of detach (size should drop to
0 but doesn't, drbd_set_my_capacity isn't called, this
is probably intentional). Loosing the size on detach
is not supported yet.

With this patch, we wait until we know a valid size for
the SCSIOP_READCAPACITY requests. The BSOD on drbdadm up
introduced recently has been fixed.