Delete the device object on IRP_MN_REMOVE_DEVICE PnP request.
This is the place to do it in the future. Here's how:

windrbd_delete_windows_device invalidates the device's
	windows_device pointer.

It then triggers a WinDRBD bus rescan.

This won't report the windows object any more.

PnP manager recognizes the removal and sends a SUPRISE_REMOVAL
	followed by a REMOVE_DEVICE PnP minor request.

Finally the device object gets deleted in REMOVE_DEVICE and
	NO_SUCH_DEVICE is returned by the pnp handler (which
	is somehow logical).

Right now, no BSOD on drbdadm down of an object that has a
valid size (others still stuck in wait_for_size(), the device
object never gets a REMOVE_DEVICE PnP request).