Reference counting in wait_for_size()
In the case a DRBD device never gets a valid size before it
is removed, the SCSI requests hang forever in wait_for_size.
We now use reference counting on waiting for size and signal
the event when the device is destroyed by the DRBD kernel
driver (via bdput()).

This patch also includes some experiments for making drbdadm
down not blue screen (it currently BSODs with the same message
as the WinAoE driver on umount).