Do not delete bus object.
And return NO_SUCH_DEVICE on SCSI if device object already deleted.
We still have a BSOD when doing a drbdadm down on an object
that never got a valid size. The reason is that REMOVE_DEVICE
is called twice, the first time it succeeds then the bus
is rescanned and then REMOVE_DEVICE is called again with
the same device object. This is probably a Windows bug
which we have to work around.