Always return STATUS_NO_SUCH_DEVICE if device object is about to be deleted.
We currently have a BSOD if the device never gets a size. It
seems that at the end of the 'open' loop REMOVE_DEVICE is sent
if there is an error, but it is sent again by PnP manager
when it discovers that the device object is gone. The first
REMOVE_DEVICE should not be sent.