Mark removed device object as dead.
We are currently setting DeviceExtention to (void*) -1 which
is in general a bad idea. Maybe one day we find a better solution.

This patch works around a Windows bug (?) that causes REMOVE_DEVICE
PnP request being sent twice. It happens when a device is removed
via drbdadm down that never got a valid size (that is never
was attached or connected).

The BSOD on drbdadm down when there was no size is now fixed.