Implementing SCSI I/O (read/write) requests.
READ requests seem to work ok, however BSOD as soon as
we initialize the disk (create an MBR using partition
manager) and probably write to the device (we don't
see it in the logs it crashes before). Also tried
becoming Primary first, same result.

Probably there is something wrong with the data transfer
address?