Copiied the address computation from WinAoE
At least the check for the mapping being valid requires more
attention. We are now able to handle requests from
user space (such as the partition manager) as well.

We should eventually unmap the mapped virtual address space
again (if this is not done by the disk driver).

Now, partition manager complains with invalid function. This
may be because we are implementing our own partition ioctl's.