Additional checks and set SrbStatus to SUCCESS on I/O
I/O via SCSI seems to work now. A new drive is created for
our WinDRBD test disk (M:) which means that one day we don't
need to take care about the drive letters any more. The
drive is currently reported as write-protected, however
the DRBD resource containg an MBR (with partitions) is
recognized and can be read/written correctly. Currently
DRBD devices can also be written when not Primary (which
is probably because there is something with drbd_open,
maybe Windows also ignores the return value of open(2)).

What is nice is that Windows takes care about the partitioning
and assignment of drive letters. We eventually can remove
that code from WinDRBD (also code for announcing drives
via the windrbd utility).