Working on printk()
We are not sure if it is responsible for some system instabilities
we observe. Most likely not (if net printk is turned off) but
the spinlocks used for memory print might also cause problems.

We will later allow the user to somehow toggle this (via ioctl
or boot parameter) and also add TCP/IP network logging. For
now we need to hardcode the printk targets into the statics
at the beginning of the file.