kmalloc_debug: log caller of alloc_page instead of alloc_page itself
And friends (alloc_page_of_size(), free_page(), free_page_kref()).
We observed lots of memory 'leaking' in DRBD, however those are
allocated for mempools which we need for proper operation even
when there are no resources up (about 64MB). They do not leak
(they will be freed when unloading the driver, which currently
does not work because of the AddDevice status).

So the reason for the system volume hangs is probably something
else than a memory leak.