Some printk's for research
Showing that Windows is accessing the SCSI port both
for reading and writing when it is idle (it shows a
panel saying new block device) it also writes to the
disk if it does not have a partition with a drive letter
assigned. So, unfortunately we cannot support auto-promote
for data drives (other that C:) because we can't tell
when Windows just writes something or the user wants
to write something.