Auto-promote on waiting for becoming primary.
We are calling try_to_promote when a boot device wants to access the
disk (right now, for all devices for testing purposes). The
reason is that boot devices are system-critical and must
be primary all the time.

We are also repeating promotion if there no access to up to date data
this is common when we are diskless and not connected yet (but
will be soon).

A follow-up patch can then remove the become primary loop in
windrbd_bootdevice.c

It seems that we can also implement a regular auto-promote: Windows
accesses the disk writable, but only at the beginning (when it
is reported via PnP).