Removed the become primary loop at the end of boot stage 2.
This is done now in the wait_for_beoming_primary function, so
whenever Windows requests something from us and we are secondary
we try to make the resource primary. Later this will happen only
for the boot device, since for data devices (other than C:) this
might not be wanted.