Always create Windows device in bdget (on creating new minor)
It was a bad idea to create it when setting the mount point.
We might change creation of the windows device later (again),
but I think for now it is good as it is. We are waiting for
size and for becoming primary in windrbd_device (upon receiving
various I/O requests), and we need early windows device creation
for the boot feature.