Fixed a BSOD introduced with 67a0d8b7
We didn't check for a valid bdev in REMOVE_DEVICE PnP request
of the disk object. Furthermore, we still need the DeviceExtension
(which is a block_device_reference) for handling the PnP
request so don't set it to NULL.

This patch also removes the deletion of the mount point. We
are not sure yet if the whole mount point thing is needed
at all and have it disabled for now.