Working on remove race.
The problem is that is the device is taken up right after taken
down a stale Windows device object may remain (because the
REMOVE_DEVICE can take some seconds until it finishes). The
clean solution would be to signal an event when REMOVE_DEVICE
finishes and remove the block_device and return from drbdadm
down only after that.