Working on getting drbdadm down work.
When doing up / down and the device never got Primary, drbdadm down
hangs forever (most likely because the REMOVE_DEVICE PnP minor
is never sent). This might be a bug in the upper layers of
Windows which we need to work around.

When the device is disabled via device manager, up/down and
up/wait-connect/down run without issues.