Debugging kref's of block_device.
Showing that kref drops to zero and is incremented thereafter
again and then drops to zero again. This seems to happen from
while waiting for REMOVE_DEVICE to be sent. Solution would be
to honor the detached_from_device_tree flag on I/O and disallow
I/O when this flag is set.