boot: Search first 640KB for DRBD ACPI table.
This is also the way Linux finds the iBFT table. We probably leave
it as it is (need to parse the drbd_config string). What currently
works is to use (again) the same driver binary for boot feature
and non-boot feature (if there is no DRBD signature, no boot device
is created and Windows will boot from the first harddisk).