First version of DRBD URI parser
Mostly opencoded but uses string helper functions whereever
possible. The format is for now a semicolon seperated string
that specifies the minimal required parameter set of connecting
to one peer (multiple peers not yet supported).

Also there is a parser_test function which is intended to be
called via a ioctl call.