Some fixes for updown loop.
.) Also check for device going down after msleep in promote
loop. This should avoid calling try_to_promote on a device
that is about to go down.

.) Always check return value of wait_for_primary. For SCSI
requests only call it when request is a WRITE request. Also
do not call it on START_DEVICE (this is probably too early).

Currently no BSOD on updown loop with wait connect, however
BSOD on updown loop when there is a peer (IRQL).

Note that there are currently lots of lots of printk's().