Started working on wait_event reimplementation.
We finally want a macro (must be a macro because of condition
expression lambda) that does everything and convenience
macros that use that central macro. This has not been
tested yet.

Unfortunately we probably have to keep the wait_event patches
(they are applied via cocci script) because I don't know how
to return a value of a macro in Microsoft Visual C.