New wait_event_timeout macro
It seems to work quite good (drbdadm up and I/O test). Now most
critical macros are ported to the new implementation. It seems
that there are no new bugs.

What really would be interresting is how Windows behaves if a
KeWaitForXXX() function is called at DISPATCH_LEVEL and if that
causes the system hangups we observe.