Not using new implementation of wait_event_interruptible
Still there is a BSOD on drbdadm up. Some wait_event function
is called outside a DRBD thread context which leads to undefined
behaviour. Either make the thread a WinDRBD thread (it seems to
be the thread created in the netlink layer) or find a general
solution to support non-WinDRBD threads.