Fixed return values of ll_wait once again.
For wait_event_interruptible_timeout we need to return the
remaining jiffies when condition evaluates to true. In addition
we need to return 1 when condition is true and the timeout elapsed.

We now have the problem that wait_event_XXX() is often also
called from a non-WinDRBD thread. We need to create task_structs
for all that threads before we proceed.

This should be done in windrbd_create_windows_thread as well as
in the entry points of windrbd_device.c