ll_wait_event_macro: break on timeout and on all errors.
What happened is that ll_wait returned -ETIMEDOUT and the
loop continued feeding -110 as timeout. This was definately
wrong. Right now there is another bug (spurious Disk failure)
which is because of wait_event_xx() called from a non-WinDRBD
thread (in drbd_md_sync_page_io(), which is run from a workqueue).