Introduced busy wait (interval 5 seconds) again.
We want to check if the NEW_CUR_UUID flag is really never
cleared in the bad case. We found that once we wait in
inc_ap_bio, the flag is not cleared thereafter under
certain circumstances. We expect this to hang also
with this timeout patch in that test case.