Removed polling loop when removing WinDRBD device.
We also observed 10 minute delays in drbdadm down (we rebooted
the machine) when polling in 10 seconds interval, so the loop is
probably not useful. Instead we must fix a certain handle leak.

drbd worker thread is still running and stuck in
wait_event_interruptible at drbd_sender.c:3008 so this is
probably the reason for REMOVE_DEVICE not being sent.