Removed old schedule function, revert to Linux interface.
Now, it is possible again to use the original Linux interface
for schedule, schedule_timeout and friends. The old schedule
function from ManTech has been commented out.

There are some extra patches to DRBD that reverts the schedule
calls to the original interface. Also we are starting on
implementing the wake_up_all (and complete_all) functions.