Fixed bug in completions.
The completion was not completed when complete() was called
before wait_for_completion(). We now have a completed flag
in the completion structure and are using wait_event_interruptible()
with the completed condition to check for completion
before we sleep.

With this commit up/down loop (without peer and without I/O)
runs for 40 iterations (then a new BSOD happened)