Auto promote only on windrbd_create requests.
We should not promote when we are about to down the device.
This should fix the need access to uptodate data on drbdadm
down. It is also much closer to the original DRBD auto-promote
mechanism (which we should use instead of our homegrown
function).

Test this with Diskless client (if the client does not BSOD
with no root device then this is ok).