Promote in SCSI_CLAIM_DEVICE
This is what we got sent on boot. We need to wait there for
Primary (but only if we are a boot device). Since Windows
does not send us a IRP_MJ_CREATE request (it might do,
but too late) this is the place for waiting for primary
for boot devices.

This patch also adds some printk's for SCSI requests.