For 'normal' (not bootdevice) devices only autopromote on create.
This appears to fix the drbdadm down takes 30 seconds and longer
bug. The problem was that when we promote a DRBD resource that
is about to go down, a lock is held and therefore PnP takes
longer to send us the REMOVE_DEVICE PnP request. Also updown
loop runs quite stable now.