Fixed 2 logic bugs
One was SCSI I/O did not work in non-boot devices (wrong condition
for else branch). The other is that we don't want to wait for size
when it is not a boot device (THINKO). The reason is that partition
manager hangs when there is no backing device and no connection.

However, now it hangs in wait_for_primary().