Experimenting with MmMapIoSpace
It fails under Windows 10 (with and without caching). This
causes DRBD not to configure the boot device which in turn
causes Windows to BSOD on boot (inaccessible boot device).

For now, we hardcode the parameters (like we did in the
beginning) until we find another solution to get the
parameters from iPXE.