drbd.cgi: Check if DRBD_MINOR parameter is numeric.
We avoid something like DRBD_MINOR=/../../etc/passwd which
would reveal the contents of /etc/passwd. drbd.cgi is now
in a state in which it can be delivered for the 0.10.0 release
a followup release should contain a fast cgi (in C) that supports
keep-alive. Reason is that it takes longer than a minute to
load all drivers (which would be much faster with a fast CGI
server).