TODO update: verifying documentation
We've found an install problem on Windows 10: there are
actually 2: one is that installing the bus driver fails
with file not found. We now have the user select the
WinDRBD application directory and install the driver from
there. The other problem is that Windows 10 (and 8) requires
a signed catalog (.cat) file for the INF. We now generate
this catalog file and install it when building releases
(building it when not building releases makes no sense
since then we are in test mode anyway ...).

We try once more if it is possible to select the driver
from the manufacturer/driver select box. If not possible
we need to change the documentation.