Creating device as DEVICE_UNKNOWN.
Right now it BSODs with 0x50 (Page fault) on drbdadm up. It is
probably a bad idea to attach the unknown device to our virtual
SCSI bus.

DEVICE_UNKNOWN devices are used by ImDisk to implement RAW devices.
There is no such thing as a Volume device (since Volumes are not
devices).