Test: Always use PnP, also for mounted devices.
Result is now that the drive letter assigned formerly with
partition manager reappears all of a sudden again. The
mount letter from the drbd.conf file does not exist now.
I/O requests are sent as SCSI requests.

I think we leave it now and talk to Phil first about how
to proceed. Using the PnP mechanism also has the benefit
that other file systems (FAT at least) also work out of the
box.

The drawback is that mounting a partition under Linux
is probably difficult (need to read the lv docs).