Create and delete dos devices
Sometimes it is good to be able to access a DRBD device
directly, we need it for example to scan for partition
(which in turn will explorer recognize the partitions).

This patch should add \DosDevices\Drbd<n> links when
creating the device (and remove them when the device
is deleted).