howto update: working on getting iSCSI working
It works now with a former WinDRBD remote boot Windows 7
machine. It shows that one can configure the boot time
IP address (and NIC) over iBFT (iPXE already does this
for iSCSI, implement it for WinDRBD as well). Furthermore
the presence of an iBFT seems also to cause network
critical boot drivers being loaded at boot time (iscsiprt
had StartType 3, BootFlags 1 and was loaded).

So good news is that the registry hacks described in
the docs and the setting of the hard coded IP address
inside the Windows VM will not be necessary any longer
(once we patch iPXE to create a iBFT entry for WinDRBD
also).