Renamed WinDRBDDriver -> WinDRBD
Windows is searching for WinDRBD (I think) Problem is that
when we load driver (via drbdadm status) then Windows starts
to search for a driver but fails. Not clear yet what it
really searches for but WinDRBD is a good guess (Bus object
is named WinDRBD). QUERY_ID seems not to be called.