Create bus device only when there is a boot config.
It should create the bus device only when booting, however
right now it creates it also when the driver is loaded at
a later point in time (e.g. loading it via windrbd install-bus-driver
in a Windows PE environment). The problem is that install-bus-driver
right now fails (also in a regular environment). We hope that
this patch fixes that.